backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "loguru>=0.7.3",
    "mcp[cli]>=1.13.1",
    "openai>=1.100.2",
    "python-dotenv>=1.1.1",
]
